Ability to link productions to prints on the print edit + rework of the link components
