+++ frontend/app/components/admin/shared/production-linker/LockedHint.vue
@@ -1,0 +1,50 @@
+<!--
+  components/admin/shared/production-linker/LockedHint.vue
+  =========================================================
+  Shown inside the ProductionLinker card when the parent entity is not saved yet.
+-->
+
+<script setup lang="ts">
+const props = defineProps<{
+  type: "blog" | "print";
+}>();
+
+const { t } = useI18n();
+
+const titleKey = computed(
+  () => `admin.${props.type}s.linkToProductionCreateHint`,
+);
+const hintKey = computed(() => `admin.${props.type}s.linkedProductionsHint`);
+</script>
+
+<template>
+  <div
+    class="flex items-start gap-3 rounded-lg border border-action-warning-border bg-action-warning-hover px-4 py-3.5"
+  >
+    <div
+      class="w-8 h-8 rounded-lg bg-action-warning-hover border border-action-warning-border flex items-center justify-center shrink-0 mt-0.5"
+    >
+      <svg
+        class="w-4 h-4 text-action-warning-icon"
+        fill="none"
+        stroke="currentColor"
+        stroke-width="2"
+        viewBox="0 0 24 24"
+      >
+        <rect x="3" y="11" width="18" height="11" rx="2" />
+        <path d="M7 11V7a5 5 0 0 1 10 0v4" stroke-linecap="round" />
+      </svg>
+    </div>
+
+    <div>
+      <p
+        class="text-[11px] font-brand font-black uppercase tracking-widest text-action-warning-icon leading-tight"
+      >
+        {{ t(titleKey) }}
+      </p>
+      <p class="mt-1 text-[10px] text-muted-foreground leading-relaxed">
+        {{ t(hintKey) }}
+      </p>
+    </div>
+  </div>
+</template>
