+++ frontend/locales/nl.json
@@ -267,7 +267,20 @@
     "of_pages": "van {total}",
     "new": "Nieuw drukwerk",
     "description": "omschrijving",
-    "relatedProductions": "Gerelateerde producties"
+    "relatedProductions": "Gerelateerde producties",
+    "linkToProduction": "Koppel aan productie",
+    "linkBtn": "Koppelen",
+    "linkToProductionSearch": "Zoek producties op titel...",
+    "linkToProductionNoResults": "Geen producties gevonden",
+    "linkToProductionCreateHint": "Sla het drukwerk eerst op om het aan een productie te koppelen.",
+    "linking": "Koppelen...",
+    "linked": "Gekoppeld",
+    "linkedProductions": "Koppel aan productie",
+    "linkedProductionsHint": "Koppel dit drukwerk aan een productie zodat het op de productiepagina verschijnt.",
+    "linkedThisSession": "Gekoppeld in deze sessie",
+    "noLinkedProductions": "Nog niet aan een productie gekoppeld",
+    "unlinkBtn": "Ontkoppelen",
+    "viewProduction": "Bekijk productiepagina"
   },
   "admin": {
     "finish": "Voltooien",
